+++ src/EncompassRest.Loans.v1/Loans/v1/ReoProperty.cs
@@ -164,7 +164,7 @@ public sealed partial class ReoProperty : DirtyExtensibleObject, IIdentifiable
     /// <summary>
     /// Mortgage Property Purchase Price [FMNN23]
     /// </summary>
-    public int? PurchasePrice { get => GetValue<int?>(); set => SetValue(value); }
+    public decimal? PurchasePrice { get => GetValue<decimal?>(); set => SetValue(value); }
 
     /// <summary>
     /// Mortgage Property Monthly Gross Rental Income [FMNN20]
